backend: refactor the pagaination config
